Moved customtodo tests to new file
